+++ fiftyone/core/runs.py
@@ -725,11 +725,13 @@ class Run(Configurable):
         if run_results is not None:
             run_results._key = None
 
-        # Must manually delete run result, which is stored via GridFS
-        if run_doc.results:
-            run_doc.results.delete()
+        if run_doc:
+            # Must manually delete run result, which is stored via GridFS
+            if run_doc.results:
+                run_doc.results.delete()
 
-        run_doc.delete()
+            run_doc.delete()
+
         dataset.save()
 
     @classmethod
